feat(projects): show the pinned instructions entry in a project chat's Files panel
Extract the pinned instructions row + editor into a shared component and reuse it
in the chat Files panel, so a project chat treats instructions.md the same way as
the project page: pinned at the top, owner-editable (read-only for others), and
filtered out of the ordinary project file list. Threads the conversation's
projectId through RightSidePanel into the panel.
