+++ platform/frontend/src/components/chat/right-side-panel.tsx
@@ -28,6 +28,9 @@ interface RightSidePanelProps {
   // Artifact props
   artifact?: string | null;
 
+  /** Set when the chat belongs to a project — enables the pinned instructions. */
+  projectId?: string | null;
+
   // Browser props
   conversationId: string | undefined;
   /** Fallback agentId for pre-conversation case */
@@ -49,6 +52,7 @@ export function RightSidePanel({
   onClose,
   canShowBrowser,
   artifact,
+  projectId,
   conversationId,
   agentId,
   onCreateConversationWithUrl,
@@ -125,6 +129,7 @@ export function RightSidePanel({
             <ConversationFilesPanel
               conversationId={conversationId}
               artifact={artifact}
+              projectId={projectId}
               onClose={onClose}
             />
           )}
